Preserve aspect ratio of eto main image
