+++ packages/web/app/components/eto/public-view/Cover.module.scss
@@ -10,12 +10,6 @@ $_logo-padding-sm: 16px;
   overflow: hidden;
   justify-content: center;
   align-items: center;
-
-  img {
-    width: 100%;
-    object-fit: cover;
-    user-select: none;
-  }
 }
 
 .company-details {
