+++ packages/web/app/components/shared/Image.module.scss
@@ -1,0 +1,8 @@
+.image {
+  // height: auto is required for browser
+  // to properly calculate image size while preserving aspect ratio
+  height: auto;
+  object-fit: cover;
+  user-select: none;
+  width: 100%;
+}
