+++ packages/web/app/components/shared/ResponsiveImage.module.scss
@@ -22,7 +22,6 @@
   @include rect(100%);
 
   display: block;
-  font-style: italic;
 }
 
 .fit-contain {
