+++ packages/web/app/components/shared/icons/TokenIcon.stories.tsx
@@ -5,4 +5,6 @@ import { TokenIcon } from "./TokenIcon";
 
 import * as tokenIcon from "../../../assets/img/token_icon.svg";
 
-storiesOf("TokenIcon", module).add("default", () => <TokenIcon srcSet={{ "1x": tokenIcon }} />);
+storiesOf("TokenIcon", module).add("default", () => (
+  <TokenIcon srcSet={{ "1x": tokenIcon }} alt="" />
+));
