+++ packages/web/app/components/shared/icons/TokenIcon.tsx
@@ -1,13 +1,13 @@
 import * as cn from "classnames";
 import * as React from "react";
 
-import { HiResImage } from "../HiResImage";
+import { Image } from "../Image";
 
 import * as styles from "./TokenIcon.module.scss";
 
-const TokenIcon: React.FunctionComponent<React.ComponentProps<typeof HiResImage>> = ({
+const TokenIcon: React.FunctionComponent<React.ComponentProps<typeof Image>> = ({
   className,
   ...props
-}) => <HiResImage {...props} className={cn(styles.tokenIcon, className)} />;
+}) => <Image {...props} className={cn(styles.tokenIcon, className)} />;
 
 export { TokenIcon };
